requirements.txt
# FastAPI base libraries
uvicorn==0.17.6
starlette==0.19.1
fastapi==0.78.0

# Handling environment variables
python-decouple==3.6

# Testing
pytest==7.1.2

# Data validation
pydantic==1.9.1
email-validator==1.2.1

# Python static typing
mypy==0.961

# ORM
SQLAlchemy==1.4.37

# PostgreSQL database adapter
psycopg2==2.9.3